+++ client/src/components/level.tsx
@@ -33,7 +33,6 @@ import { DeletedChatContext, InputModeContext, PreferencesContext, MonacoEditorC
 import { DualEditor } from './infoview/main'
 import { GameHint, InteractiveGoalsWithHints, ProofState } from './infoview/rpc_api'
 import { DeletedHints, Hint, Hints, MoreHelpButton, filterHints } from './hints'
-import { PrivacyPolicyPopup } from './popup/privacy_policy'
 import path from 'path';
 
 import '@fontsource/roboto/300.css'
@@ -73,18 +72,15 @@ function Level() {
   const gameInfo = useGetGameInfoQuery({game: gameId})
 
   // pop-ups
-  const [privacy, setPrivacy] = React.useState(false)
   const [preferencesPopup, setPreferencesPopup] = React.useState(false)
 
-  function closePrivacy()   {setPrivacy(false)}
   function closePreferencesPopup() {setPreferencesPopup(false)}
   function togglePreferencesPopup() {setPreferencesPopup(!preferencesPopup)}
 
   return <WorldLevelIdContext.Provider value={{worldId, levelId}}>
     {levelId == 0 ?
-      <Introduction privacy={privacy} setPrivacy={setPrivacy} togglePreferencesPopup={togglePreferencesPopup} /> :
-      <PlayableLevel key={`${worldId}/${levelId}`} privacy={privacy} setPrivacy={setPrivacy} togglePreferencesPopup={togglePreferencesPopup}/>}
-    {privacy ? <PrivacyPolicyPopup handleClose={closePrivacy} /> : null}
+      <Introduction togglePreferencesPopup={togglePreferencesPopup} /> :
+      <PlayableLevel key={`${worldId}/${levelId}`}  togglePreferencesPopup={togglePreferencesPopup}/>}
     {preferencesPopup ? <PreferencesPopup handleClose={closePreferencesPopup} /> : null}
   </WorldLevelIdContext.Provider>
 }
@@ -213,7 +209,7 @@ function ExercisePanel({codeviewRef, visible=true}: {codeviewRef: React.MutableR
   </div>
 }
 
-function PlayableLevel({ privacy, setPrivacy, togglePreferencesPopup }) {
+function PlayableLevel({ togglePreferencesPopup }) {
   let { t } = useTranslation()
   const codeviewRef = useRef<HTMLDivElement>(null)
   const gameId = React.useContext(GameIdContext)
@@ -249,8 +245,6 @@ function PlayableLevel({ privacy, setPrivacy, togglePreferencesPopup }) {
   const [lockEditorMode, setLockEditorMode] = useState(false)
   const [typewriterInput, setTypewriterInput] = useState("")
   const lastLevel = levelId >= gameInfo.data?.worldSize[worldId]
-
-  function togglePrivacy() {setPrivacy(!privacy)}
 
   // When clicking on an inventory item, the inventory is overlayed by the item's doc.
   // If this state is set to a pair `(name, type)` then the according doc will be open.
@@ -419,7 +413,6 @@ function PlayableLevel({ privacy, setPrivacy, togglePreferencesPopup }) {
                   isLoading={level.isLoading}
                   levelTitle={(mobile ? "" : t("Level")) + ` ${levelId} / ${gameInfo.data?.worldSize[worldId]}` +
                     (level?.data?.title && ` : ${t(level?.data?.title, {ns: gameId})}`)}
-                  togglePrivacy={togglePrivacy}
                   togglePreferencesPopup={togglePreferencesPopup}
                   />
                 {mobile?
@@ -483,7 +476,7 @@ function IntroductionPanel({gameInfo}) {
 export default Level
 
 /** The site with the introduction text of a world */
-function Introduction({privacy, setPrivacy, togglePreferencesPopup}) {
+function Introduction({ togglePreferencesPopup}) {
   let { t } = useTranslation()
 
   const gameId = React.useContext(GameIdContext)
@@ -497,11 +490,8 @@ function Introduction({privacy, setPrivacy, togglePreferencesPopup}) {
 
   let image: string = gameInfo.data?.worlds.nodes[worldId].image
 
-  const togglePrivacy = () => {
-    setPrivacy(!privacy)
-  }
   return <>
-    <LevelAppBar isLoading={gameInfo.isLoading} levelTitle={t("Introduction")} togglePrivacy={togglePrivacy} togglePreferencesPopup={togglePreferencesPopup}/>
+    <LevelAppBar isLoading={gameInfo.isLoading} levelTitle={t("Introduction")} togglePreferencesPopup={togglePreferencesPopup}/>
     {gameInfo.isLoading ?
       <div className="app-content loading"><CircularProgress /></div>
     : mobile ?
