+++ client/src/components/welcome.tsx
@@ -13,7 +13,6 @@ import { Button } from './button'
 import { PreferencesContext } from './infoview/context'
 import { InventoryPanel } from './inventory'
 import { ErasePopup } from './popup/erase'
-import { PrivacyPolicyPopup } from './popup/privacy_policy'
 import { RulesHelpPopup } from './popup/rules_help'
 import { UploadPopup } from './popup/upload'
 import { PreferencesPopup} from "./popup/preferences"
@@ -84,21 +83,16 @@ function Welcome() {
 
   // pop-ups
   const [eraseMenu, setEraseMenu] = React.useState(false)
-  const [privacy, setPrivacy] = React.useState(false)
   const [info, setInfo] = React.useState(false)
   const [rulesHelp, setRulesHelp] = React.useState(false)
   const [uploadMenu, setUploadMenu] = React.useState(false)
   const [preferencesPopup, setPreferencesPopup] = React.useState(false)
 
   function closeEraseMenu()   {setEraseMenu(false)}
-  function closePrivacy()   {setPrivacy(false)}
-  function closeInfo()        {setInfo(false)}
   function closeRulesHelp()   {setRulesHelp(false)}
   function closeUploadMenu()  {setUploadMenu(false)}
   function closePreferencesPopup() {setPreferencesPopup(false)}
   function toggleEraseMenu()  {setEraseMenu(!eraseMenu)}
-  function togglePrivacy()  {setPrivacy(!privacy)}
-  function toggleInfo()       {setInfo(!info)}
   function toggleUploadMenu() {setUploadMenu(!uploadMenu)}
   function togglePreferencesPopup() {setPreferencesPopup(!preferencesPopup)}
 
@@ -114,9 +108,8 @@ function Welcome() {
       <CircularProgress />
     </Box>
   : <>
-    <WelcomeAppBar pageNumber={pageNumber} setPageNumber={setPageNumber} gameInfo={gameInfo.data} togglePrivacy={togglePrivacy}
-      toggleEraseMenu={toggleEraseMenu} toggleUploadMenu={toggleUploadMenu}
-      toggleInfo={toggleInfo} togglePreferencesPopup={togglePreferencesPopup}/>
+    <WelcomeAppBar pageNumber={pageNumber} setPageNumber={setPageNumber} gameInfo={gameInfo.data}
+      toggleEraseMenu={toggleEraseMenu} toggleUploadMenu={toggleUploadMenu} togglePreferencesPopup={togglePreferencesPopup}/>
     <div className="app-content">
       { mobile ?
           <div className="welcome mobile">
@@ -138,7 +131,6 @@ function Welcome() {
           </Split>
       }
     </div>
-    {privacy ? <PrivacyPolicyPopup handleClose={closePrivacy} /> : null}
     {rulesHelp ? <RulesHelpPopup handleClose={closeRulesHelp} /> : null}
     {eraseMenu? <ErasePopup handleClose={closeEraseMenu}/> : null}
     {uploadMenu? <UploadPopup handleClose={closeUploadMenu}/> : null}
